+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added `sys` internal module and `sys.path` to manage module paths
 - changed error raiser to show file and line number of code in error rendering
+- changed import command to import modules without quotes. for example `import @hash`. stil you can use quotes.
 - fixed an bug in module bootstraper, now the `PASHMAKPATH` environment variable has default empty value
 
 ## 0.4 (2020-12-1)